Fix UI and add a logo
The UI uses now the ttk widgets instead of the default tkinter widgets.
This change was made because on macOS there is bug that on some
machines the text on the buttons disappears. Since the client uses
apples OS, this has priority.

Other than that there is temporary logo and many build scipts got added
to freeze the code.